requirement.txt
Django==2.2.7
django-crispy-forms==1.8.1
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==6.2.1
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
